Condensar print y añadir plato a comida
Se han condensado las funcionalidades print, además de mostrar continuamente el procesamiento del alimento (crudo/cocinado).

Por otro lado, se muestran los valores nutricionales del plato actual, que posteriormente se añade a la comida al indicar que se quiere incluir uno nuevo. El problema es que al hacer esto, los valores que ahora se muestran están a 0 porque son los del nuevo plato aún vacío.
Se podría hacer mostrando la información de la comida directamente en lugar del plato, pero entonces habría que añadir el plato a la comida desde el primer momento e ir actualizando la comida conforme se añaden ingredientes al plato, no solo al querer añadir otro.

¿Podría mostrar los valores del plato actual en "Comida actual" y los de la comida en "Acumulado hoy"? ¿O eso dista demasiado del funcionamiento establecido desde un principio en el que solo se modifica "Acumulado hoy" al guardar la comida?

+++ src/smartcloth/Ingrediente.h
@@ -9,20 +9,20 @@
 class Ingrediente{
   private:
     Grupo _grupo;
-    float _peso;
+    double _peso;
     ValoresNutricionales _valoresIng;
 
   public:
     Ingrediente();
-    Ingrediente(Grupo grupo, float peso); 
+    Ingrediente(Grupo grupo, double peso); 
     
     inline void setGrupoIng(Grupo grupo){ _grupo = grupo; };
     inline Grupo getGrupoIng(){ return _grupo; };
     
-    inline void setPesoIng(float peso){ _peso = peso; };
-    inline float getPesoIng(){ return _peso; };
+    inline void setPesoIng(double peso){ _peso = peso; };
+    inline double getPesoIng(){ return _peso; };
 
-    void updateValoresIng(Grupo grupo, float peso);
+    void updateValoresIng(Grupo grupo, double peso);
     inline void setValoresIng(ValoresNutricionales val){ _valoresIng.setValores(val); };
     inline ValoresNutricionales getValoresIng(){ return _valoresIng; };
 };
@@ -34,18 +34,18 @@ Ingrediente::Ingrediente(){
     this->setPesoIng(0.0);
 }
 
-Ingrediente::Ingrediente(Grupo grupo, float peso){
+Ingrediente::Ingrediente(Grupo grupo, double peso){
     this->setGrupoIng(grupo);
     this->setPesoIng(peso);
     this->updateValoresIng(grupo, peso);
 }
 
 
-void Ingrediente::updateValoresIng(Grupo grupo, float peso){
-    float carb = grupo.Carb_g * peso;
-    float lip = grupo.Lip_g * peso;
-    float prot = grupo.Prot_g * peso;
-    float kcal = grupo.Kcal_g * peso;
+void Ingrediente::updateValoresIng(Grupo grupo, double peso){
+    double carb = grupo.Carb_g * peso;
+    double lip = grupo.Lip_g * peso;
+    double prot = grupo.Prot_g * peso;
+    double kcal = grupo.Kcal_g * peso;
     ValoresNutricionales valAux(carb, lip, prot, kcal);
     _valoresIng.setValores(valAux);
 }
